requirements.txt
torch==2.1.0
torchvision==0.16.0
pennylane==0.33.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1
tqdm==4.66.1
psutil==5.9.6
